test: move polling srv records prose tests to one file
