+++ test/unit/sdam/srv_polling.test.js
@@ -10,8 +10,6 @@ const dns = require('dns');
 const EventEmitter = require('events').EventEmitter;
 const chai = require('chai');
 const sinon = require('sinon');
-const mock = require('../../tools/mongodb-mock/index');
-const { HostAddress, isHello } = require('../../../src/utils');
 const { MongoDriverError } = require('../../../src/error');
 
 const expect = chai.expect;
@@ -292,138 +290,6 @@ describe('Mongos SRV Polling', function () {
       expect(srvPoller.start).to.have.been.calledOnce;
       emit(topologyDescriptions[2], topologyDescriptions[3]);
       expect(srvPoller.start).to.have.been.calledOnce;
-    });
-
-    describe('prose tests', function () {
-      function srvAddresses(records) {
-        return records.map(r => `${r.name}:${r.port}`);
-      }
-
-      const MONGOS_HELLO = Object.assign({}, mock.HELLO, {
-        msg: 'isdbgrid'
-      });
-
-      beforeEach(function () {
-        return Promise.all(Array.from({ length: 4 }).map(() => mock.createServer())).then(
-          servers => {
-            context.servers = servers;
-          }
-        );
-      });
-
-      afterEach(function () {
-        return mock.cleanup();
-      });
-
-      afterEach(function (done) {
-        if (context.topology) {
-          context.topology.close(done);
-        } else {
-          done();
-        }
-      });
-
-      function runSrvPollerTest(recordSets, done) {
-        context.servers.forEach(server => {
-          server.setMessageHandler(request => {
-            const doc = request.document;
-
-            if (isHello(doc)) {
-              request.reply(Object.assign({}, MONGOS_HELLO));
-            }
-          });
-        });
-
-        const srvPoller = new FakeSrvPoller({ srvHost: SRV_HOST });
-        const seedlist = recordSets[0].map(record =>
-          HostAddress.fromString(`${record.name}:${record.port}`)
-        );
-
-        context.topology = new Topology(seedlist, { srvPoller, srvHost: SRV_HOST });
-        const topology = context.topology;
-
-        topology.connect({}, err => {
-          if (err) {
-            return done(err);
-          }
-
-          try {
-            expect(topology.description).to.have.property('type', TopologyType.Sharded);
-            const servers = Array.from(topology.description.servers.keys());
-
-            expect(servers).to.deep.equal(srvAddresses(recordSets[0]));
-
-            topology.once('topologyDescriptionChanged', function () {
-              tryDone(done, function () {
-                const servers = Array.from(topology.description.servers.keys());
-
-                expect(servers).to.deep.equal(srvAddresses(recordSets[1]));
-              });
-            });
-
-            process.nextTick(() => srvPoller.trigger(recordSets[1]));
-          } catch (e) {
-            done(e);
-          }
-        });
-      }
-
-      // The addition of a new DNS record:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27019  localhost.test.build.10gen.cc.
-      it('should handle the addition of a new DNS record', function (done) {
-        const recordSets = [
-          [srvRecord(context.servers[0]), srvRecord(context.servers[1])],
-          [
-            srvRecord(context.servers[0]),
-            srvRecord(context.servers[1]),
-            srvRecord(context.servers[2])
-          ]
-        ];
-        runSrvPollerTest(recordSets, done);
-      });
-
-      // The removal of an existing DNS record:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27018  localhost.test.build.10gen.cc.
-      it('should handle the removal of an existing DNS record', function (done) {
-        const recordSets = [
-          [srvRecord(context.servers[0]), srvRecord(context.servers[1])],
-          [srvRecord(context.servers[0])]
-        ];
-        runSrvPollerTest(recordSets, done);
-      });
-
-      // The replacement of a DNS record:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27018  localhost.test.build.10gen.cc.
-      // replace by:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27019  localhost.test.build.10gen.cc.
-      it('should handle the replacement of a DNS record', function (done) {
-        const recordSets = [
-          [srvRecord(context.servers[0]), srvRecord(context.servers[1])],
-          [srvRecord(context.servers[0]), srvRecord(context.servers[2])]
-        ];
-        runSrvPollerTest(recordSets, done);
-      });
-
-      // The replacement of both existing DNS records with one new record:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27019  localhost.test.build.10gen.cc.
-      it('should handle the replacement of both existing DNS records with one new record', function (done) {
-        const recordSets = [
-          [srvRecord(context.servers[0]), srvRecord(context.servers[1])],
-          [srvRecord(context.servers[2])]
-        ];
-        runSrvPollerTest(recordSets, done);
-      });
-
-      // The replacement of both existing DNS records with two new records:
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27019  localhost.test.build.10gen.cc.
-      // _mongodb._tcp.test1.test.build.10gen.cc.  86400  IN SRV  27020  localhost.test.build.10gen.cc.
-      it('should handle the replacement of both existing DNS records with two new records', function (done) {
-        const recordSets = [
-          [srvRecord(context.servers[0]), srvRecord(context.servers[1])],
-          [srvRecord(context.servers[2]), srvRecord(context.servers[3])]
-        ];
-        runSrvPollerTest(recordSets, done);
-      });
     });
   });
 });
